FlaskORFastApi/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
python-dotenv>=0.19 # For loading .env files
gunicorn>=20.0 # Optional: For production deployment 